Convert ids to long. discord snowflakes are larger than ints. Added API_URL for code for oauth. Made API key environment aware (master vs prod)

+++ src/main/java/com/greatmancode/legendarybotapi/discorduser/DiscordUserBackend.java
@@ -4,7 +4,7 @@ import com.greatmancode.legendarybotapi.impl.aws.dynamodb.DynamoDBHelper;
 
 public class DiscordUserBackend {
 
-    public static DiscordUser getDiscordUser(int id) {
+    public static DiscordUser getDiscordUser(long id) {
         if (System.getenv("AWS_EXECUTION_ENV") != null) {
             return DynamoDBHelper.getDiscordUser(id);
         }
